refactor: mypgae 모델 레이어 분리 및 쿼리키 컨벤션 적용
