+++ src/widgets/mypage/model/mypage.types.ts
@@ -1,0 +1,6 @@
+export type TabValue = 'all' | 'favorite' | 'created';
+
+export interface TabList {
+  value: TabValue;
+  label: string;
+}
